% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posterior_Predictive_p_Function.R
\name{posterior.p.value}
\alias{posterior.p.value}
\title{Posterior predictive p-value function for the DROIIDS Model}
\usage{
posterior.p.value(analysis.name = NA, seed = NA)
}
\arguments{
\item{analysis.name}{A character string giving the name of the analysis that was used in the \code{DROIIDS} function.}

\item{seed}{The specified seed for generating random numbers.}
}
\description{
This function produces a posterior predictive p-value assessing goodness of fit
using the posterior estimates from data analyzed with the \code{DROIIDS} function.
}
\details{
The posterior iterations (\code{paste("DROIIDS.", <analysis.name>, ".Iterations", sep="")}) in the global environment is
   used to calculate a posterior predictive p-value.  The \code{posterior.p.value} function will automatically produce this
   dataset if it does not already exist in the global environment.

   For each iteration in the \code{DROIIDS} function output, predicted mean outcome values are calculated using the
   observed iteration estimates for \eqn{a_{i,t}} and \eqn{\beta} with the data model (\eqn{z.mean_{i,t,l}=\Phi[l,] a_{i,t} +
   X_{i} \beta}). Potential outcome values are generated from a normal distribution with the
   predicted mean outcome values as the mean and respective iteration location variance, \eqn{\sigma^2_{\epsilon,l}}.
   That is, \eqn{z.gen_{i,t,l} ~ N(z.mean_{i,t,l}, \sigma^2_{\epsilon,l})}. This occurs for every iteration across all chains.

   A goodness of fit calculation is calculated for the generated values, \eqn{z.gen_{i,t,l}}, and the observed values,
    \eqn{z.obs_{i,t,l}}, using the following equations for person \eqn{i}, time \eqn{t}, and location \eqn{l}:
   \deqn{fit.z.gen = \Sigma_{i} \Sigma_{t} \Sigma_{l} ((z.gen_{i,t,l} - z.mean_{i,t,l})^2 / \sigma^2_{\epsilon,l})}
   \deqn{fit.z.obs = \Sigma_{i} \Sigma_{t} \Sigma_{l} ((z.obs_{i,t,l} - z.mean_{i,t,l})^2 / \sigma^2_{\epsilon,l})}

   Using these fit values, the proportion of times (fit.z.gen > fit.z) was calculated across all iterations and chains.
   When the model fits the data well, this proportion should be near 0.5.  Since this calculation involves observed data,
   the posterior predictive p-value cannot be computed when data are imputed.

   This function will return the posterior predictive p-value in the global environment with the name
   \code{paste("DROIIDS.", <analysis.name>, ".posterior.p", sep="")}.

   The \code{analysis.name} is the character string that was specified in the \code{DROIIDS} function.  This analysis name will
   link the \code{posterior.p.value} function to the iterations that were saved both in the global environments
   and in .txt files in your current working directory.  If the iteration data are not currently in the global
   environment, this function will automatically read in all the necessary information before calculating the posterior predictive
   p-value.
}

