% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rates_Change_Function.R
\name{rates.of.change}
\alias{rates.of.change}
\title{Rates of change function for the DROIIDS Model}
\usage{
rates.of.change(analysis.name = NULL, save = FALSE, return.rate = TRUE)
}
\arguments{
\item{analysis.name}{A character string giving the name of the analysis that was used in the \code{DROIIDS} function.}

\item{save}{An indicator (\code{TRUE}/\code{FALSE}) of whether or not a CSV file should be created in the current working directory
recording rates of change for each individual. By default, the rates are saved in the global environment as
\code{paste("DROIIDS.", <analysis.name>, ".rates", sep="")}.}

\item{return.rate}{An indicator (\code{TRUE}/\code{FALSE}) of whether or not the rates of change should be returned
and printed.}
}
\description{
Using the predicted estimates that were created through the \code{prediction} function in the DROIIDS
package, this function calculates the rates of change in time across individuals and locations.
}
\details{
A linear regression line is fit to the predicted data using time as the explanatory variable for every location
   across all individuals.  This provides an estimate of the linear rate of change.  Using the predicted data instead
   of the raw data accounts for measurement variability in the location values.  The \code{prediction} function
   will automatically be performed to produce the predicted data if the \code{paste("DROIIDS.", <analysis.name>, ".predicted"
   , sep="")} dataset does not already exist in the global environment.

   The \code{analysis.name} is the character string that was specified in the \code{DROIIDS} function.  This analysis name will
   link the \code{rates.of.change} function to the iterations and output that were saved both in the global environments
   and in .txt files in your current working directory.

   \code{save = TRUE} will produce a CSV file in the current working directory with rates of change for individuals
   at all locations. This provides rate estimates regardless of whether there are missing values in the raw data. The output file will
   be named \code{paste("DROIIDS_", <analysis.name>, "_Rates_of_Change.csv", sep="")}.
}

